Move agent test notification execution logic to view model
